Upgrade terraform-provider-commons and add slug creation logic.
Upgraded the version of 'terraform-provider-commons' from v1.0.2 to v1.0.3 in 'go.mod' and 'go.sum'. Additionally, a 'createSlug' modifier, that generates a slug from repository name, has been added to the 'resource_bitbucket_repository.go'. When the name of the repository changes, it triggers resource recreation in Terraform. This update provides more precise handling of attribute changes.